Add files via upload
The proposed system involves the use of congestion free routers which are basically edge routers that are placed at the end of a network which patrol the flow’s packet in order to avoid congestion collapse and inequitable bandwidth allocation. If the incoming packets are more than the outgoing packets it means there is buffering or packets are being discarded. This congestion collapse prevention comes at a cost as there will be some complexity at the edge routers as they are monitoring the rates of individual flow. The complexity can be reduced with TCP friendly rate control.